Added the new package of OOP exercises and a README
