+++ src/OOP/exerciseEight/Main.java
@@ -1,4 +1,4 @@
-package exerciseEight;
+package OOP.exerciseEight;
 /*Write a Java program to create class called "TrafficLight"
  with attributes for color and duration,
  and methods to change the color and check for red or green.
