+++ src/OOP/exerciseSix/Main.java
@@ -1,4 +1,4 @@
-package exerciseSix;
+package OOP.exerciseSix;
 /*Write a Java program to create a class called "Employee"
 with a name, job title, and salary attributes,
 and methods to calculate and update salary by a given percentage.
