+++ src/README.md
@@ -1,0 +1,31 @@
+# Java Concepts Exercises Repository
+
+Welcome to my Java Concepts Repository! This repository serves as a collection of exercises I have done to have a better and deeper understanding of Java as a programming language.
+
+## Table of Contents
+
+- OOP
+- Exception Handling
+- File Input/Output
+
+## Structure
+
+The repository is organized into folders based on different Java concepts and topics. Each folder contains exercises, sample code, and notes. And some of these are:
+
+- **Exception Handling**: Understanding try-catch blocks, custom exceptions, etc.
+- **File I/O**: Reading and writing files, handling streams, etc.
+- **OOP Principles**: Encapsulation, inheritance, polymorphism, etc.
+- **Collections Framework**: Lists, maps, sets, and their operations.
+
+## Instructions on how to run the program
+1. **Clone the repository:**
+   ```bash
+   git clone https://github.com/RominaCaballero/Java-Concepts-Exercises.git
+   ```
+2. Open the project in IntelliJ IDEA:
+
+- Open IntelliJ IDEA and select **File > Open**.
+- Navigate to the folder where the repository is cloned and open the project.
+
+3. Choose which concept you are interested and which number of exercise you wanna run.
+4. Run the program you chose and see it come to life!
